Move tests to integration library
