+++ test/Datadog.Trace.Tests/DiagnosticListeners/AspNetCoreDiagnosticObserverTests.cs
@@ -7,12 +7,13 @@ using Datadog.Trace.Agent;
 using Datadog.Trace.Configuration;
 using Datadog.Trace.DiagnosticListeners;
 using Datadog.Trace.Sampling;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-#if NETCOREAPP2_1
 using Microsoft.AspNetCore.Hosting.Internal;
-#endif
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using Xunit;
 using Xunit.Sdk;
@@ -21,7 +22,7 @@ namespace Datadog.Trace.Tests.DiagnosticListeners
 {
     [Collection(nameof(TracerInstanceTestCollection))]
     [TracerRestorer]
-    public class AspNetCoreDiagnosticObserverAmbientTracerTests
+    public class AspNetCoreDiagnosticObserverTests
     {
         [Fact]
         public async Task<string> CompleteDiagnosticObserverTest()
@@ -29,7 +30,7 @@ namespace Datadog.Trace.Tests.DiagnosticListeners
             Tracer.Instance = GetTracer();
 
             var builder = new WebHostBuilder()
-                .UseStartup<MvcStartup>();
+                .UseStartup<Startup>();
 
             var testServer = new TestServer(builder);
             var client = testServer.CreateClient();
@@ -59,11 +60,7 @@ namespace Datadog.Trace.Tests.DiagnosticListeners
 
             IObserver<KeyValuePair<string, object>> observer = new AspNetCoreDiagnosticObserver(tracer);
 
-#if NETCOREAPP2_1
             var context = new HostingApplication.Context { HttpContext = GetHttpContext() };
-#else
-            var context = new { HttpContext = GetHttpContext() };
-#endif
 
             observer.OnNext(new KeyValuePair<string, object>("Microsoft.AspNetCore.Hosting.HttpRequestIn.Start", context));
 
@@ -110,6 +107,19 @@ namespace Datadog.Trace.Tests.DiagnosticListeners
             return httpContext;
         }
 
+        private class Startup
+        {
+            public void ConfigureServices(IServiceCollection services)
+            {
+                services.AddMvc();
+            }
+
+            public void Configure(IApplicationBuilder builder)
+            {
+                builder.UseMvcWithDefaultRoute();
+            }
+        }
+
         [AttributeUsage(AttributeTargets.Class, Inherited = true)]
         private class TracerRestorerAttribute : BeforeAfterTestAttribute
         {
@@ -126,6 +136,24 @@ namespace Datadog.Trace.Tests.DiagnosticListeners
                 Tracer.Instance = _tracer;
                 base.After(methodUnderTest);
             }
+        }
+    }
+
+    /// <summary>
+    /// Simple controller used for the aspnetcore test
+    /// </summary>
+#pragma warning disable SA1402 // File may only contain a single class
+    public class HomeController : Controller
+    {
+        public async Task<string> Index()
+        {
+            await Task.Yield();
+            return "Hello world";
+        }
+
+        public void Error()
+        {
+            throw new Exception();
         }
     }
 }
